+++ .vscode/settings.json
@@ -20,4 +20,7 @@
     "rust-analyzer.linkedProjects": [
         "scripts/magma_search/Cargo.toml",
     ],
+    "[rust]": {
+        "editor.tabSize": 4,
+    }
 }
